main/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
finnhub-python==2.4.20


# beautifulsoup4==4.13.4
# certifi==2025.4.26
# charset-normalizer==3.4.2
# click==8.1.8
# colorama==0.4.6
# feedparser==6.0.11
# filelock==3.18.0
# fsspec==2025.3.2
# huggingface-hub==0.30.2
# idna==3.10
# Jinja2==3.1.6
# joblib==1.5.0
# MarkupSafe==3.0.2
# mpmath==1.3.0
# networkx==3.4.2
# nltk==3.9.1
# numpy==2.2.5
# packaging==25.0
# PyYAML==6.0.2
# regex==2024.11.6
# requests==2.32.3
# safetensors==0.5.3
# sgmllib3k==1.0.0
# soupsieve==2.7
# sympy==1.14.0
# textblob==0.19.0
# tokenizers==0.21.1
# torch==2.7.0
# tqdm==4.67.1
# transformers==4.51.3
# typing_extensions==4.13.2
# urllib3==2.4.0